# openapi.yaml
openapi: 3.0.0
info:
  title: MindPal API
  description: Complete documentation of all API endpoints for the MindPal Mental Health Application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /profile:
    post:
      tags: [Profile]
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                personalInfo:
                  type: object
                  properties:
                    firstName:
                      type: string
                      example: John
                    lastName:
                      type: string
                      example: Doe
                    dateOfBirth:
                      type: string
                      format: date
                      example: "1995-06-15"
                    gender:
                      type: string
                      enum: [male, female, other, prefer-not-to-say]
                      example: male
                    profilePicture:
                      type: string
                      example: "https://example.com/photo.jpg"
                contactInfo:
                  type: object
                  properties:
                    email:
                      type: string
                      example: john@example.com
                    phone:
                      type: string
                      example: "+1234567890"
                    address:
                      type: object
                      properties:
                        street:
                          type: string
                          example: "123 Main St"
                        city:
                          type: string
                          example: "New York"
                        state:
                          type: string
                          example: "NY"
                        country:
                          type: string
                          example: "US"
                        zipCode:
                          type: string
                          example: "10001"
                preferences:
                  type: object
                  properties:
                    theme:
                      type: string
                      enum: [light, dark, system]
                      example: system
                    language:
                      type: string
                      example: en
                    notificationPreferences:
                      type: object
                      properties:
                        email:
                          type: boolean
                          example: true
                        push:
                          type: boolean
                          example: true
                        sms:
                          type: boolean
                          example: false
                healthInfo:
                  type: object
                  properties:
                    height:
                      type: number
                      description: Height in cm
                      example: 175
                    weight:
                      type: number
                      description: Weight in kg
                      example: 70
                    bloodType:
                      type: string
                      enum: ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"]
                      example: "O+"
                    allergies:
                      type: array
                      items:
                        type: string
                      example: ["peanuts"]
                    medications:
                      type: array
                      items:
                        type: string
                      example: []
                    conditions:
                      type: array
                      items:
                        type: string
                      example: []
      responses:
        "201":
          description: Profile created

    get:
      tags: [Profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile details

    put:
      tags: [Profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Same structure as create profile - all fields are optional for update
              properties:
                personalInfo:
                  type: object
                  properties:
                    firstName:
                      type: string
                      example: Jane
                    lastName:
                      type: string
                      example: Doe
                preferences:
                  type: object
                  properties:
                    theme:
                      type: string
                      enum: [light, dark, system]
                      example: dark
      responses:
        "200":
          description: Profile updated

    delete:
      tags: [Profile]
      summary: Delete profile
      responses:
        "200":
          description: Profile deleted
...
  /voice/upload:
    post:
      tags: [Voice]
      summary: Upload voice note for transcription
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (wav, webm, mp3, ogg, m4a)
                sessionId:
                  type: string
                  description: Optional. Valid MongoDB ObjectId of an existing chat session.
                  example: ""
                language:
                  type: string
                  description: Language code for transcription
                  example: en-US
                generateAudio:
                  type: string
                  description: Set to "true" to generate TTS audio response
                  enum: ["true", "false"]
                  example: "false"
      responses:
        "200":
          description: Transcription successful
  /voice/transcribe:
    post:
      tags: [Voice]
      summary: Transcribe audio only
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio]
              properties:
                audio:
                  type: string
                  format: binary
                  description: Audio file (wav, webm, mp3, ogg, m4a)
                language:
                  type: string
                  description: Language code for transcription
                  example: en-US
      responses:
        "200":
          description: Transcription result
  /voice/synthesize:
    post:
      tags: [Voice]
      summary: Convert text to speech
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: Hello, how are you feeling today?
                voice:
                  type: string
                  description: Voice ID (ElevenLabs or Google voice name)
                  example: en-US-Neural2-F
                speed:
                  type: number
                  description: Speaking rate (0.5 to 2.0)
                  example: 1.0
      responses:
        "200":
          description: Audio content
  /voice/analyze/{sessionId}:
    get:
      tags: [Voice]
      summary: Analyze voice patterns
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
            description: A valid MongoDB ObjectId for the session
            example: 60d0fe4f5311236168a109ca
      responses:
        "200":
          description: Analysis result

  # Mood
...